requirements.txt
torch==1.7.0
transformers==4.4.2
textworld==1.4.0
tqdm==4.60.0
numpy==1.19.1
python-Levenshtein==0.12.2